test/e2e: move operator checks at prepare.go
